Implement leave_file so cases can leave their file in place if requested to
git-svn-id: https://svn.domain51.com/test/trunk@81 54d25d7b-d40e-47af-a99d-8d4037bc02b6

+++ src/Domain51/Test/Case.php
@@ -4,6 +4,7 @@ class Domain51_Test_Case
 {
     public $name = 'This is a sample test case to show that "Hello World" can be echoed';
     public $filename = '';
+    public $leave_file = false;
     
     public function __construct($name, $filename, $code, $sections)
     {
@@ -17,7 +18,9 @@ class Domain51_Test_Case
     
     public function __destruct()
     {
-        unlink($this->filename);
+        if ($this->leave_file == false) {
+            unlink($this->filename);
+        }
     }
     
     public function update()
